sentinel/requirements.txt
djangorestframework==3.11.1
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-language==2.0.0